Update aBitAdvancedCalculator.cpp
Updated the code and fixed a bug, at line 47 the user input was mishandled. It was because the input_num2 was 
used for input twice instead of one. So when chosing the option N whatever you will enter the code will print out 
wrong calculated value. But i fixed it now :)